requirements.txt
google-genai>=0.1.0
moviepy>=2.1.1
pydub>=0.25.1
streamlit>=1.41.1